Refactoring: Move functions from Evaluator to EvaluationStrategy (#447)
