+++ src/main/haskell/kore/src/Kore/Step/Function/Evaluator.hs
@@ -8,22 +8,12 @@ Stability   : experimental
 Portability : portable
 -}
 module Kore.Step.Function.Evaluator
-    ( applyFirstSimplifierThatWorks
-    , evaluateApplication
-    -- TODO(virgil): Move all exports except evaluateApplication to
-    -- EvaluationStrategy.hs.
-    , evaluateWithDefinitionAxioms
-    , AcceptsMultipleResults (..)
-    , evaluateBuiltin
+    ( evaluateApplication
     ) where
 
 import           Control.Exception
                  ( assert )
-import           Control.Monad
-                 ( when )
 import qualified Data.Map as Map
-import           Data.Maybe
-                 ( isJust )
 import           Data.Reflection
                  ( give )
 import qualified Data.Text as Text
@@ -35,34 +25,21 @@ import           Kore.IndexedModule.MetadataTools
                  ( MetadataTools (..) )
 import           Kore.Predicate.Predicate
                  ( makeTruePredicate )
-import           Kore.Step.AxiomPatterns
-                 ( EqualityRule (EqualityRule) )
-import           Kore.Step.BaseStep
-                 ( OrStepResult (OrStepResult),
-                 UnificationProcedure (UnificationProcedure),
-                 stepWithRemaindersForUnifier )
-import qualified Kore.Step.BaseStep as OrStepResult
-                 ( OrStepResult (..) )
 import           Kore.Step.ExpandedPattern
                  ( ExpandedPattern, PredicateSubstitution, Predicated (..) )
-import qualified Kore.Step.ExpandedPattern as ExpandedPattern
-                 ( fromPurePattern )
 import           Kore.Step.Function.Data
                  ( AttemptedAxiomResults (AttemptedAxiomResults),
                  BuiltinAndAxiomSimplifier (..), BuiltinAndAxiomSimplifierMap )
 import           Kore.Step.Function.Data as AttemptedAxiom
                  ( AttemptedAxiom (..) )
-import           Kore.Step.Function.Data as AttemptedAxiomResults
+import qualified Kore.Step.Function.Data as AttemptedAxiomResults
                  ( AttemptedAxiomResults (..) )
-import           Kore.Step.Function.Matcher
-                 ( unificationWithAppMatchOnTop )
 import qualified Kore.Step.Merging.OrOfExpandedPattern as OrOfExpandedPattern
                  ( mergeWithPredicateSubstitution )
 import           Kore.Step.OrOfExpandedPattern
                  ( OrOfExpandedPattern )
 import qualified Kore.Step.OrOfExpandedPattern as OrOfExpandedPattern
-                 ( extractPatterns, flatten, isFalse, make, merge,
-                 traverseWithPairs )
+                 ( flatten, make, merge, traverseWithPairs )
 import           Kore.Step.Pattern
 import           Kore.Step.Simplification.Data
                  ( PredicateSubstitutionSimplifier, SimplificationProof (..),
@@ -72,17 +49,6 @@ import           Kore.Step.StepperAttributes
                  ( Hook (..), StepperAttributes (..), isSortInjection_ )
 import           Kore.Unparser
 import           Kore.Variables.Fresh
-
-{-|Describes whether simplifiers are allowed to return multiple results or not.
--}
-data AcceptsMultipleResults = WithMultipleResults | OnlyOneResult
-    deriving (Eq, Ord, Show)
-
-{-|Converts 'AcceptsMultipleResults' to Bool.
--}
-acceptsMultipleResults :: AcceptsMultipleResults -> Bool
-acceptsMultipleResults WithMultipleResults = True
-acceptsMultipleResults OnlyOneResult = False
 
 {-| 'evaluateApplication' - evaluates functions on an application pattern.
 -}
@@ -207,201 +173,41 @@ evaluateApplication
                     simplifier
                     patt
 
-evaluateBuiltin
-    :: forall variable level
-    .   ( FreshVariable variable
-        , MetaOrObject level
-        , Ord (variable level)
-        , OrdMetaOrObject variable
-        , SortedVariable variable
-        , Show (variable level)
-        , Show (variable Object)
-        , Unparse (variable level)
-        , ShowMetaOrObject variable
-        )
-    => BuiltinAndAxiomSimplifier level
-    -> MetadataTools level StepperAttributes
-    -> PredicateSubstitutionSimplifier level Simplifier
-    -> StepPatternSimplifier level variable
-    -> StepPattern level variable
+
+evaluateSortInjection
+    :: (MetaOrObject level, Ord (variable level))
+    => MetadataTools level StepperAttributes
+    -> OrOfExpandedPattern level variable
+    -> Application level (StepPattern level variable)
     -> Simplifier
-        (AttemptedAxiom level variable, SimplificationProof level)
-evaluateBuiltin
-    (BuiltinAndAxiomSimplifier builtinEvaluator)
-    tools
-    substitutionSimplifier
-    simplifier
-    patt
-  = do
-    (result, _proof) <-
-        builtinEvaluator
-            tools
-            substitutionSimplifier
-            simplifier
-            patt
-    case result of
-        AttemptedAxiom.NotApplicable
-          | isPattConcrete
-          , Just hook <- getAppHookString ->
-            error
-                (   "Expecting hook " ++ hook
-                ++  " to reduce concrete pattern\n\t"
-                ++ show patt
+        (OrOfExpandedPattern level variable, SimplificationProof level)
+evaluateSortInjection tools unchanged ap = case apChild of
+    (App_ apHeadChild grandChildren)
+      | give tools isSortInjection_ apHeadChild ->
+        let
+            [fromSort', toSort'] = symbolOrAliasParams apHeadChild
+            apHeadNew = updateSortInjectionSource apHead fromSort'
+        in
+            assert (toSort' == fromSort) $
+            return
+                ( OrOfExpandedPattern.make
+                    [ Predicated
+                        { term = apHeadNew grandChildren
+                        , predicate = makeTruePredicate
+                        , substitution = mempty
+                        }
+                    ]
+                , SimplificationProof
                 )
-          | otherwise ->
-            return (AttemptedAxiom.NotApplicable, SimplificationProof)
-        AttemptedAxiom.Applied _ -> return (result, SimplificationProof)
+    _ -> return (unchanged, SimplificationProof)
   where
-    isPattConcrete = traceFunction D_Function_evaluateApplication [] $ isJust (asConcretePurePattern patt)
-    appHead = case patt of
-        App_ head0 _children -> head0
-        _ -> error
-            ("Expected an application pattern, but got " ++ show patt ++ ".")
-    -- TODO(virgil): Send this from outside after replacing `These` as a
-    -- representation for application evaluators.
-    getAppHookString = traceFunction D_Function_evaluateApplication [] $
-        Text.unpack <$> (getHook . hook . symAttributes tools) appHead
-
-applyFirstSimplifierThatWorks
-    :: forall variable level
-    .   ( FreshVariable variable
-        , MetaOrObject level
-        , Ord (variable level)
-        , OrdMetaOrObject variable
-        , SortedVariable variable
-        , Show (variable level)
-        , Show (variable Object)
-        , Unparse (variable level)
-        , ShowMetaOrObject variable
-        )
-    => [BuiltinAndAxiomSimplifier level]
-    -> AcceptsMultipleResults
-    -> MetadataTools level StepperAttributes
-    -> PredicateSubstitutionSimplifier level Simplifier
-    -> StepPatternSimplifier level variable
-    -> StepPattern level variable
-    -> Simplifier
-        (AttemptedAxiom level variable, SimplificationProof level)
-applyFirstSimplifierThatWorks [] _ _ _ _ _ =
-    return
-        ( AttemptedAxiom.NotApplicable
-        , SimplificationProof
-        )
-applyFirstSimplifierThatWorks
-    (BuiltinAndAxiomSimplifier evaluator : evaluators)
-    multipleResults
-    tools
-    substitutionSimplifier
-    simplifier
-    patt
-  = do
-    (applicationResult, _proof) <-
-        evaluator tools substitutionSimplifier simplifier patt
-
-    case applicationResult of
-        AttemptedAxiom.Applied AttemptedAxiomResults
-            { results = orResults
-            , remainders = orRemainders
-            } -> do
-                when
-                    (length (OrOfExpandedPattern.extractPatterns orResults) > 1
-                    && not (acceptsMultipleResults multipleResults)
-                    )
-                    -- We should only allow multiple simplification results
-                    -- when they are created by unification splitting the
-                    -- configuration.
-                    -- However, right now, we shouldn't be able to get more
-                    -- than one result, so we throw an error.
-                    (error
-                        (  "Unexpected simplification result with more "
-                        ++ "than one configuration: "
-                        ++ show applicationResult
-                        )
-                    )
-                when
-                    (not (OrOfExpandedPattern.isFalse orRemainders)
-                    && not (acceptsMultipleResults multipleResults)
-                    )
-                    -- It's not obvious that we should accept simplifications
-                    -- that change only a part of the configuration, since
-                    -- that will probably make things more complicated.
-                    --
-                    -- Until we have a clear example that this can actually
-                    -- happen, we throw an error.
-                    (error
-                        (  "Unexpected simplification result with remainder: "
-                        ++ show applicationResult
-                        )
-                    )
-                return (applicationResult, SimplificationProof)
-        AttemptedAxiom.NotApplicable ->
-            applyFirstSimplifierThatWorks
-                evaluators
-                multipleResults
-                tools
-                substitutionSimplifier
-                simplifier
-                patt
-
-evaluateWithDefinitionAxioms
-    :: forall variable level
-    .   ( FreshVariable variable
-        , MetaOrObject level
-        , Ord (variable level)
-        , OrdMetaOrObject variable
-        , SortedVariable variable
-        , Show (variable level)
-        , Show (variable Object)
-        , Unparse (variable level)
-        , ShowMetaOrObject variable
-        )
-    => [EqualityRule level]
-    -> MetadataTools level StepperAttributes
-    -> PredicateSubstitutionSimplifier level Simplifier
-    -> StepPatternSimplifier level variable
-    -> StepPattern level variable
-    -> Simplifier
-        (AttemptedAxiom level variable, SimplificationProof level)
-evaluateWithDefinitionAxioms
-    definitionRules tools substitutionSimplifier simplifier patt
-  = do
-    let
-        expanded :: ExpandedPattern level variable
-        expanded = ExpandedPattern.fromPurePattern patt
-
-    (OrStepResult { rewrittenPattern, remainder }, _proof) <-
-        stepWithRemaindersForUnifier
-            tools
-            (UnificationProcedure unificationWithAppMatchOnTop)
-            substitutionSimplifier
-            (map (\ (EqualityRule rule) -> rule) definitionRules)
-            expanded
-    let
-        remainderResults :: [ExpandedPattern level variable]
-        remainderResults = OrOfExpandedPattern.extractPatterns remainder
-
-        simplifyPredicate
-            :: ExpandedPattern level variable
-            -> Simplifier
-                (ExpandedPattern level variable, SimplificationProof level)
-        simplifyPredicate =
-            ExpandedPattern.simplifyPredicate
-                tools
-                substitutionSimplifier
-                simplifier
-
-    simplifiedRemainderList <- mapM simplifyPredicate remainderResults
-    let
-        simplifiedRemainderResults :: [ExpandedPattern level variable]
-        (simplifiedRemainderResults, _proofs) =
-            unzip simplifiedRemainderList
-    return
-        ( AttemptedAxiom.Applied AttemptedAxiomResults
-            { results = rewrittenPattern
-            , remainders = OrOfExpandedPattern.make simplifiedRemainderResults
-            }
-        , SimplificationProof
-        )
+    apHead = applicationSymbolOrAlias ap
+    [fromSort, _] = symbolOrAliasParams apHead
+    [apChild] = applicationChildren ap
+    updateSortInjectionSource head1 fromSort1 =
+        mkApp toSort1 head1 { symbolOrAliasParams = [fromSort1, toSort1] }
+      where
+        [_, toSort1] = symbolOrAliasParams head1
 
 {-| 'reevaluateFunctions' re-evaluates functions after a user-defined function
 was evaluated.
@@ -456,42 +262,7 @@ reevaluateFunctions
             mergedPatt
     return evaluatedPatt
 
-evaluateSortInjection
-    :: (MetaOrObject level, Ord (variable level))
-    => MetadataTools level StepperAttributes
-    -> OrOfExpandedPattern level variable
-    -> Application level (StepPattern level variable)
-    -> Simplifier (OrOfExpandedPattern level variable, SimplificationProof level)
-evaluateSortInjection tools unchanged ap = case apChild of
-    (App_ apHeadChild grandChildren)
-      | give tools isSortInjection_ apHeadChild ->
-        let
-            [fromSort', toSort'] = symbolOrAliasParams apHeadChild
-            apHeadNew = updateSortInjectionSource apHead fromSort'
-        in
-            assert (toSort' == fromSort) $
-            return
-                ( OrOfExpandedPattern.make
-                    [ Predicated
-                        { term = apHeadNew grandChildren
-                        , predicate = makeTruePredicate
-                        , substitution = mempty
-                        }
-                    ]
-                , SimplificationProof
-                )
-    _ -> return (unchanged, SimplificationProof)
-  where
-    apHead = applicationSymbolOrAlias ap
-    [fromSort, _] = symbolOrAliasParams apHead
-    [apChild] = applicationChildren ap
-    updateSortInjectionSource head1 fromSort1 =
-        mkApp toSort1 head1 { symbolOrAliasParams = [fromSort1, toSort1] }
-      where
-        [_, toSort1] = symbolOrAliasParams head1
-
-{-| 'mergeWithCondition' ands the given condition-substitution to the given
-function evaluation.
+{-| Ands the given condition-substitution to the given function evaluation.
 -}
 mergeWithConditionAndSubstitution
     ::  ( MetaOrObject level
